OpenAI-API/TrainingAnswerer/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyautogui>=0.9.53
Pillow>=10.0.0
pywinstyles>=1.0.0
pyinstaller>=6.0.0
pytesseract>=0.3.10
